renaming makemWin.js to makem.js and cleaning the code
